requirements.txt
selenium==4.5.0
webdriver-manager==4.0.1
google-cloud-bigquery
google-cloud-storage
beautifulsoup4==4.11.1
pandas==1.5.1
numpy==1.23.4
db-dtypes==1.0.4
lxml==4.9.1
undetected-chromedriver==3.4.6
requests==2.31.0
PyVirtualDisplay==3.0
fsspec==2022.8.2
gcsfs==2022.8.2
openpyxl==3.0.10
selenium-wire==5.1.0
xlrd==1.2.0
gspread==5.7.2
kafka-python==2.0.2
timeout-decorator==0.5.0
Pillow==10.0.0
opencv-python==4.8.0.74
sewar==0.4.5
fake-useragent==1.4.0
blinker<1.8.0
instaloader==4.11
cryptography==39.0.1
pytz==2022.5
bitwardentools==1.0.57